+++ src/main/java/cc/cassian/raspberry/ModCompat.java
@@ -32,4 +32,5 @@ public final class ModCompat {
     public static final boolean BLUEPRINT = ModList.get().isLoaded("blueprint");
     public static final boolean BREWINANDCHEWIN = ModList.get().isLoaded("brewinandchewin");
     public static final boolean EMI = ModList.get().isLoaded("emi");
+    public static final boolean CONTROLLABLE = ModList.get().isLoaded("controllable");
 }