+++ src/main/java/cc/cassian/raspberry/compat/controllable/ControllableCompat.java
@@ -1,0 +1,9 @@
+package cc.cassian.raspberry.compat.controllable;
+
+import com.mrcrayfish.controllable.client.ButtonBindings;
+
+public class ControllableCompat {
+    public static boolean isJumping() {
+         return ButtonBindings.JUMP.isButtonDown();
+    }
+}